Restructuring database for easier development and starting to use jinja
