+++ run.py
@@ -1,0 +1,5 @@
+from game2048 import app
+
+# Use Ctrl + Shift + R to hard refresh browser
+if __name__ == "__main__":
+    app.run(debug=True)
